requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
tqdm>=4.60.0
concurrent-log-handler>=0.9.20
pandas>=2.0.0
pdfplumber>=0.10.0
tabula-py==2.7.0
pandas==2.0.3
